requirements.txt
azure-cognitiveservices-vision-customvision==1.0.0
azure-common==1.1.23
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
docutils==0.15.2
future==0.18.2
futures==3.1.1
geocoder==1.38.1
idna==2.8
imageio==2.6.1
isodate==0.6.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msrest==0.6.10
networkx==2.4
numpy==1.17.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
Pillow==6.2.1
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.4
python-dateutil==2.8.1
PyWavelets==1.1.1
ratelim==0.1.6
requests==2.22.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scipy==1.3.1
six==1.12.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.38.0
urllib3==1.25.6
virtualenv==16.7.5
wincertstore==0.2
